+++ R/traits_sex.R
@@ -1,0 +1,30 @@
+#' @title Length, Width, Weight, and Sex traits dataframe.
+#' @description Length, Width, Weight, and Sex traits dataframe.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: Length, Width, Weight, and Sex traits dataframe (1st-col: Value, 2nd-col: Traits, 3rd-col: Sex).
+#' @usage data(traits_sex)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/BoxStat/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(traits_sex)
+#'
+#' # 3. View example data
+#' traits_sex
+#'
+"traits_sex"
+
+# data <- read.table(file = "data-tables/traits_sex.txt",
+# 									 header = TRUE,
+# 									 sep = "\t",
+# 									 stringsAsFactors = F)
+# traits_sex <- data
+# usethis::use_data(traits_sex)