+++ R/weight_sex.R
@@ -1,0 +1,30 @@
+#' @title Weight and Sex traits dataframe.
+#' @description Weight and Sex traits dataframe.
+#' @author benben-miao
+#'
+#' @docType data
+#' @format Dataframe: Weight and Sex traits dataframe (1st-col: Weight, 2nd-col: Sex).
+#' @usage data(weight_sex)
+#'
+#' @keywords datasets
+#'
+#' @references https://github.com/BioSciTools/BioSciToolsDatasets/tree/main/QuantileQuantile/
+#'
+#' @examples
+#' # 1. Library TOmicsVis package
+#' library(TOmicsVis)
+#'
+#' # 2. Load example data
+#' data(weight_sex)
+#'
+#' # 3. View example data
+#' weight_sex
+#'
+"weight_sex"
+
+# data <- read.table(file = "data-tables/weight_sex.txt",
+# 									 header = TRUE,
+# 									 sep = "\t",
+# 									 stringsAsFactors = F)
+# weight_sex <- data
+# usethis::use_data(weight_sex, overwrite = TRUE)